third_party/WebKit/LayoutTests/fast/dom/empty-anchor-in-overflow-scroller-expected.txt

PASS scroller.scrollTop is 20
PASS successfullyParsed is true

TEST COMPLETE

text
